Added initial PotionTypes and EntityTypes implementation.
In order to avoid duplicate entity names with mods, all entities will be stored in an internal registry as 'modid:name'. If you call EntityType.getByName(id)
and do not include the modid, it will assume 'minecraft:name'.

Potions currently do not have this feature but I plan on adding if needed.

* Added MixIn for EntityRegistry in order to catch mod entities and add
them to our internal registry.