Bug fix for ServerTickEvent error
The ServerTickError was assumed to have
a Phase.START and Phase.END timing.  In test,
the counter was twice as slow as expected.
Testing without discrimination of Phase yields
test results that are closer to expected. There's
still a high factor of latency (30%) to debug and
a non linear slew (5-28%).