SyncScheduler Bug fixes
Fixed some bugs in counting ticks.  Phase.START is the leading
edge and each leading edge (more or less occurs at 50ms intervals).
There is a measured and expected latency every few milliseconds
and the server rushes to catch up. Noted in the timestamps of
each ServerTickEvent.  The features added to the API for Optional<T>
return values are implemented.