Implement suggested changes
Lots of changes were required to implement everything.

Implementing getGeneratorPopulators turned out to be a challenge. The solution is to inject some code in the Chunk constructor that takes a ChunkPrimer. Unfortunatily, by the time the mixin is called, the blocks from the ChunkPrimer have already been moved to the chunk. This means that the blocks in the chunk have to be changed, not the blocks in the ChunkPrimer.

I made some changes to how the populators of a world are accessed: no longer through world.getGenerator().accessPopulators, but through ((IMixinWorld)world).getPopulators().

All the buffers have been reorganized a bit.