Implement world creation. Closes #218
In order to generate a proper level.dat per world, every world will now
use a seperate save handler unlike Forge/Vanilla which use only one.
This is required in order to have world settings saved per world.

To create a custom world, you will be required to use our new
WorldBuilder. This builder provides you will all the settings you will
need in order to create a world. If there are settings you want to modify
post creation, use our WorldProperties interface.

Note: When creating a world, it will NOT be loaded by default. The world
folder is created along with initial data. If you want the world to be
available, you must load it by calling loadWorld. However, if you call
our build() method in WorldBuilder, we will attempt to create and load
the world.

Upon starting server, Sponge will now scan for available worlds and
register them with Forge's DimensionManager in order to load properly. If
a world was created with 'enabled' or 'loadOnStartup' set to false, it
will not be loaded at startup. This should help plugins such as MultiVerse
which had to support this themselves.

UUID support has been added and is now stored with the world data.
TeleportHelper has been implemented to support safe teleports.

All of our custom world data as well as plugin data will be stored in a
new level dat file called "level_sponge.dat". We chose to store all
additional world data in a seperate file to avoid loss of data if a user
launched the server in a non-forge or sponge environment.