Add support for modded clients when using IP forwarding.
This (at the moment) requires a special build of BungeeCord built from
https://github.com/dualspiral/BungeeCord/commit/a03cf4e08ff131d7e1fd031fbb38691e17c4f2ec

This build adds support for the FML marker added by Forge during
the initial handshake.
This build will currently NOT work with non-Sponge servers.

Until this change is merged into Bungee proper, a build may be found at
http://downloads.drnaylor.co.uk/BungeeCord-FML.jar