More fun work on event refactor.
* Start work on converting sponge event data for Forge
* Add support for Subscribe's beforeModifications.
Plugins will now be able to subscribe before events are passed to forge
handlers. If you wish to subscribe before and after, you need to use two
handlers.

The event flow is as follows :

1. Plugin handlers are checked for beforeModifications set to true,
event is fired to all matching handlers.
2. Plugin event data is converted for Forge and sent to forge handlers.
3. Plugin handlers are checked for beforeModifications set to false
(default) and fired to matching handlers.

* Add initial support for mods calling Dimension.initDimension.
Note: This is still untested.
* Allow plugins to alter replacement block during PlayerPlaceBlockEvent.