Sync event data from plugins before passing to Forge and mod handlers.
Due to the way Sponge handles events, there are many Mixins where we
inject new data that is stored seperately from Forge. Not all of
this data has a corresponding setter and must be synced properly in order
to make sure both Forge and mod handlers receive the correct data.

* Sync event data from Forge to plugin listeners that do not have
beforeModifications set to true.

In order to understand the event flow and syncing better, here is what
happens after a Forge event is fired :

1. Forge event gets passed to Sponge Event bus.
2. Sponge passes event to any plugin listener that has beforeModifications
set to true.
3. Sponge syncs plugin data from step 2 then passes to mod handlers.
4. Sponge syncs forge data from step 3 then passes to plugin listeners
that have beforeModifications set to false which is the default.
5. Sponge passes event result back to Forge event bus.