Refactor event integration between Forge and Sponge. Fixes #342, Fixes #343
Due to the way Forge handles PlayerInteractEvent, a new approach had to be
made in order to fire off seperate events to Sponge. A new method
'createSpongeEvent' has been added to the base FML event in order to create
a sponge event based on criteria.

For example, Forge's PlayerInteractEvent handles all InteractionTypes
within the same event. In order to fire the appropriate Primary and Secondary
events to Sponge, the event mixin overrides createSpongeEvent and returns the
appropriate event based on Forge's Action type.

* Add onForgePost method to SpongeForgeEventFactory in order to handle special
cases such as ExplosionEvent.Detonate.
* Fix client crash generating BlockSnapshot in PlayerInteractEvent.
* Update SpongeCommon ref.