Ignore block state captures for existing blocks during PlaceBlockEvent.
In 1.8, mojang removed World's setBlockMetadataWithNotify method in favor
of the new BlockStates. Currently, all block changes are now passed to World's
setBlockState method. Due to these changes, our block capture functionality
began to receive block state changes for replaced blocks causing our capture
list to grow unexpected.

One of the main issues was with chests. After placing a single chest down,
the block capture would record 2 block snapshots with the first being what was
replaced such as an air block and the second being the state change of the
chest due to a direction change. To workaround this issue, we now only
capture if the replaced block is different than the existing block. This
essentially ignores state change captures for existing blocks during a
capture.