Implement cause tracking for SpawnEntityEvent and DropEntityEvent.
* All entity items dropped into the world due to a death or destruction
will now be passed in 'DropItemEvent.Destruct'. Cancelling this event will
prevent the items from being dropped into the world.

* All entity spawns into the world will now be tracked by
SpawnEntityEvent. If a spawn is caused by the vanilla spawner, it will be
passed to 'SpawnEntityEvent.Spawner' first. If a spawn is caused by a
chunk loading it will be passed to 'SpawnEntityEvent.ChunkLoad'. Keep in
mind that these events now send entities in bulk which means you may
receive 1 or more. Any spawns not tracked internally will be passed to
'SpawnEntityEvent.Custom'.

* Removed DropItemEvent.Harvest.
  - This event is now handled by DropItemEvent.Destruct
* Added support for spawnEntity cause from API.
* Added support for sending SpawnEntityEvent's for bulk entities.
  - Forge is still handled 1 entity at a time.
* Added tracking for child entities.
  - Example: If you spawn a wolf and the wolf kills a skeleton, it will have
    you in the cause.
* Fixed User sometimes being sent as cause for players that were online.
* Track last item used by players when placing or interacting with a
block.
* Track commands changing blocks or spawning entities.