Force close screen on client if InteractBlockEvent is cancelled.
Some mods such as OpenComputers open GUI's on client-side only after a
block interaction. If the InteractBlockEvent.Secondary is cancelled, the
GUI will remain open on client since the server has no idea about the
container changing. To workaround this, we will always send a close window
packet to guarantee the client has nothing open if event is cancelled.

* Clean up some code in MixinPlayerInteractionManager.