Fix Forge's BlockEvent.PlaceEvent never firing. Fixes #1024
As for the BreakEvent, our capturing system records all block changes in
the world as they happen then fires events to mods at end. This causes
getBlockSnapshot to return AIR since the block has already been changed at
that point. In order to workaround this, you will need to construct a fresh
BlockSnapshot using the constructor with the data passed in event.