Re-factor handling of Forge's BreakEvent. Fixes #857
Forge treats their BreakEvent as a Pre event BEFORE the break actually occurs
while Sponge fires a break event AFTER it occurs in order to capture.
This causes mods to receive the BreakEvent AFTER HarvestDropsEvent which
breaks mods such as EnderIO's BlockPoweredSpawner.

To workaround this issue, the BreakEvent will only be fired in
PlayerInteractionManager#tryHarvest along with a 'ChangeBlockEvent.Pre'
so plugins have a chance to alter the final result.