Don't restore player item on cancel if set to dead. Fixes #1048
Some mods such as Flux-Networks kill the entity item being dropped in
order to spawn a "custom" one. The only way to determine if we should
not restore is for the mod to set the entity to dead. Luckily, Flux-Networks
does this so we can validate the cancel and restore if needed.

The proper fix would be to PR Forge and allow mods to replace the entity
being spawned during EntityJoinWorldEvent. For now this workaround should
be good enough until a proper fix is handled in Forge.