Move MixinWorld out of the 'server' config section
This section is only for classes that physically exist on only the
dedicated client or server (e.g. 'Minecraft', 'DedicatedServer', etc)
or classes that we only want to modify on a dedicated side (e.g. the FML
marker in 'C00Handshake')

We want to mix into World in singleplayer, so this mixin needs to go in
the normal config section