Update SpongeCommon for ExtraUtilities 2 fixes.
* Call EntityJoinWorld event in all dropped item cases.

Forge mods always expect an EntityJoinWorld event to fire when an item spawns
into the world.

* Add missing MixinDimensionType to json. This should solve mod dimensions
not loading on startup even when set to do so.

Fixes #1368, Fixes #1378, Fixes #1410
Fixes #1440, Fixes #1466, Fixes #1468
Fixes #1480, Fixes #1488

Note: Due to an issue with world configs not properly saving loadOnStartup
properly, you will need to regenerate your ExtraUtilities 2 sponge config
folder which can be found under ./config/sponge/worlds/extrautils2. Simply
delete the folder and upon restart, the folder should be generated
properly. Make sure the DIM-9999 world config's loadOnStartup is set to
true as the mod requires the dimension be loaded on startup. Do the same
for any other mods that expect to load on startup but are not.