Move SpongeCoremod out of java6 source set
On Minecraft 1.12+, checking for Java versions older than Java 8
does no longer work in the coremod, because other Java 8 classes
are loaded before our coremod class is loaded.

Additionally, having the coremod separately in the java6 source set
causes issues with IntelliJ IDEA 2017.2.

We can just move it back to the main source set, but still keep the
Java 8 version check because it also verifies that users are at least
on Java 8u40.