Properly handle ticking items with Forge
Call the Forge-added method Item#onUsingTick from the newly created
SpongeCommon hook. Previously, we were skipping this method entirely

Also, ensure that cancelling the Forge LivingEntityUseItemEvent.Tick
causes the Sponge event to be updated properly. Cancelling the Forge
event causes the player to stop using the item entirely, but cancelling
the Sponge event simply causes that tick's actions to be skipped.
Therefore, we manually set the duration of the Sponge event to -1 if the
Forge event is cancelled.

Fixes #2247