Provide player to ChannelRegistrationEvent.
This event fires on the Netty thread in Forge and we cannot delay a tick to fire it on the main thread (as a plugin dev may want to quickly unregister a channel).

Remove pointless nethandler context as we cannot make a frame here. Instead, provide the Platform type being registered on. I would appreciate feedback here.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>