Fix server thread check on dedicated client
On the dedicated client, we still run things on the server thread when
the server is 'almost' shut down - specifically, firing the
ServerStoppedEvent. To ensure that the CauseStackManager doesn't
erroneously detect us as being off the main thread, we only null out the
cached thread when the server thread is about to exit.