Add Integration Tests for Schematics.
This performs an actual Schematic placement is tested for the full
circle deserialization, placement, recreation, serialization, and
deserialization. The test iteslf needs to be expanded to perform other
integration tests, such as version 1 schematics to version 2, but
overall it will succeed in validating that Schematics themselves will
work.

Fixes #2785.

Merges #2791.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>