Implement ChangeBlockEvent.Pre for modded liquids. Fixes #892.
A little explanation about this addition. For starters, we now modify
the Forge added base classes for custom liquids, but as many know, we
cannot mix into other mod code for implementors of IFluidBlock that do
not extend BlockFluidBase/Classic/Finite etc. This remains fully
implemented for the cases of mod blocks that do extend those classes,
and not override the methods themselves. Due to the nature of how the
injections work however, a bug with Mixin's refmap generation was
discovered, leading to a required addition to the legacy-long-forgotten
extraSrg.srg file. It now contains a searge mapping for both
BlockFluidFinite and BlockFluidClassic #updateTick method since
previously, the refmap would point to
net/minecraft/block/Block;updateTick().

This has been minimally tested with some mod fluids and a sponge audit
to ensure that mixin application succeeds.

To keep up to date with the Mixin issue, it can be referenced here:
SpongePowered/Mixin#326. As soon as the issue is resolved, the
extraSrg.srg file can be removed from the project.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>