Add constraints for Forge build version.
In relation to #2805, because some server modpacks are shipping an older
version of Forge, and SpongeForge is building against a newer version
compared to that, our injections to Forge specific classes ended up
failing because of the target method signature changes. This adds better
constraint logging on our injector when attempting to run on older
versions of Forge than 2821. Note that the issue is not a bug in Sponge
since the issue itself is server owners running an outdated version of
Forge that SpongeForge is simply not backwards compatible with.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>