Compatibility changes for World Generation pipeline. Fixes #2582.
The key here with these changes are to allow forge mods having their
own custom IChunkGenerator to be provided by the ChunkProvider, while
still giving Sponge complete control of the world generation pipeline.

The overall change here is that unlike vanilla, Sponge will not re-assign
the chunk generator on the provider, but simply redirect access to the
generation in the desired call sites for generation. This allows for mods
that need to check the generator type to do so, thereby fixing compatibility
with some features that mods depend on having their customized provider.

Fixes #2582.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>