Support mods who extend Teleporter and override placeEntity.
Let me preface this with the fact that this is not supported by Forge
(they added the interface to be used for this exact reason).

This change delegates the placeEntity call entirely on the mod's behalf.

This change also fixes mods who rely on the client receiving an unknown
dimension id to provider id and bails to the Overworld.

Closes #2931.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>