Bump common for final method changes.
Due to the issues with mods unintentionally and potentially
overriding a SpongeAPI mixed in method, there's no verification
or "exception handler" for us to hook in to when the verify
error occurs. Because of this, it will be an expected error
case for the mod author to potentially change their method
name to be compatible with Sponge. Natively, the methods
are declared final through SpongeCommon onto the specific
targets, so only during start up, when the mods content
classes are being read and verified, will the VerifyError
be thrown.

This resolves SpongePowered/SpongeCommon#1845
but with the caveat that the error is now much more
apparent.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>