Update SpongeCommon for various fixes.
Note: there is an important fix that may cause custom data to not be retained when a player respawns (e.g. on death).
Plugins that use custom data should check to see if they are impacted, if so, they should copy data on the RespawnPlayerEvent
using the `getOriginalPlayer` and `getPlayer` calls to get the original player to get the data from and the new player to
add the data to, respectively.