Pass a derived world info to Forge's WorldEvent.Load.
Many Forge mods understand and operate solely on dimension ids for
worlds. Some expect that a World's name in the WorldInfo WILL be the
actual world's name. This is not true for Vanilla sub-levels, those use
a DerivedWorldInfo and inherit all of it off of the Vanilla level
"overworld". This fakes it so that the derived version we pass to this
event will have the world name set.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>